Added confirm to Pet

+++ frontend/src/components/Pet.jsx
@@ -17,10 +17,11 @@ const Pet = () => {
   const currentPet = petData.filter((pet) => pet._id === petID)
 
   const handlePetDelete = () => {
-    if (currentPet[0]._id === petID) {
-      dispatch(deletePet(petID))
-      navigate('/pets')
-
+    if (window.confirm('Are you sur you want to delete this pet?')) {
+      if (currentPet[0]._id === petID) {
+        dispatch(deletePet(petID))
+        navigate('/pets')
+      }
     }
   }
